Added some comments to the build config files before I forget what the options do.

+++ Apps/edit/Classes/ECCodeView.m
@@ -887,7 +887,7 @@ static inline CGFloat square_distance(CGPoint a, CGPoint b)
     if (closest < lineRange.length - 1)
     {
         lineRange = CTLineGetStringRange(line);
-        if (result == lineRange.location + lineRange.length)
+        if ((long)result == lineRange.location + lineRange.length)
             result--;
     }
